requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
nltk>=3.8.1
requests>=2.31.0
pdfminer.six>=20221105
reportlab>=4.0.0
beautifulsoup4>=4.12.0